go.mod: Use `require` for runc and containderd, the comment is wrong
Go will use the highest of all the versions mentioned in a require; if
we upgraded to a Helm that still has the old versions of runc and
containerd, it would still use the newer versions mentioned in our
`require() // indirect` block.

That's actually what the purpose of indirect `require`s is; to force a
newer version than what your dependencies want.  You should only need
to use a `replace` to force an _older_ version.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>