Prevent deadlock when connectorWorker encounters problem early.
If the connector encountered problems when connecting to the daemon
or to the cluster, the connector process would sometimes lock up
forever waiting for the cluster and traffic-manager to become ready.
Now, a nil entry will be posted on those channels when a failure occurs
so that they are unblocked.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>