Add Tunnel and WatchDial to gRPC.
A Tunnel represents one single connection where the client or
traffic-agent represents one end (the client side) and the
traffic-manager represents the other (the server side). The first
message that a client sends when the tunnel is established is will
always contain the session ID, connection ID, and timeouts used by
the dialer endpoints.

`WatchDial` makes it possible for the client side to receive
`DialRequest`s from the `traffic-manager`. Requests are sent when an
intercepted `traffic-agent` creates a `Tunnel` that needs to be extended
to the Telepresence client on the workstation, or the other way around,
when that client creates a tunnel that needs to be extended to an
intercepted `traffic agent`. The receiver of the request dials a
connection and responds with a `Tunnel`. The manager then connects the
two tunnels.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>