Don't redirect log output to Tempdir() in telepresence_test
There's really no reason to use a special directory for the logfiles
just because the test checks what it contains. A new logfile is created
on each run anyway. The logfiles are normally gathered at the end of the
run and using a temporary directory prevents that. Furthermore, there's
a fair chance that a daemon process isn't quite dead yet when the tests
end which often yields weird cleanup errors when deleting the tempdir
(especially on Windows).

Signed-off-by: Thomas Hallgren <thomas@datawire.io>