Don't let cloud API key retrieval slow down the remain calls.
A slow response from SystemA must not have a negative effect on the
client's call to `manager.Remain()`. This commit decouples one from
the other.

The time until a session is considered stale is also increased from 15
to 30 seconds because calls can get have a high latency on a busy
network.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>