Give annotation drop some time to take effect.
The `TestP_SuccessfullyInterceptsHeadlessService` often fails with an
"object has been modified" error in the "injected from command" test.
The reason for this is that the "injected from webhook" test removes
an injection annotation which creates a raise condition between the
completion of that action and the "injected from command" test's attempt
to modify the same workload.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>