Remove optimization of DNS update.
Just picking the last update of search paths might result in
an attempt to set the exact same search paths as the previous
setting. That in turn is a no-op. This is bad if the paths were
changed in between since it actually disables the DNS-flush
on some systems (notably darwin).

Signed-off-by: Thomas Hallgren <thomas@datawire.io>