Make no attempt to remove pods that don't exist or are in conflict.
The `refreshReplicaSet()` function makes an attempt to remove all pods
owned by a replicaset, which is fine. But the list it acts no may
contain pods that have been deleted already (the deletion just hasn't
finished yet) which means that at the time of the new attempt to delete
the pod, it may be in conflict or even not exist anymore. This commit
inserts a check to avoid a new deletion attempt when this happens.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>