Fix token refresh issue
The telepresence login command creates a newTokenSource with the context
of the command, which is canceled once the command finishes. The
tokenSource is stateful and used by the refresh goroutine to refresh the
token, so now we pass the token to that gorotuine so it can re-set the
tokenSource once the login command is complete and the refresh functions
since that context hasn't expired.

Signed-off-by: Donny Yung <donaldyung@datawire.io>