Drop support for mux tunnel connections
This commit removes almost all traces of the mux tunnel used in
Telepresence version 2.4.4 and earlier which means backward
compatibility with versions <= 2.4.4 is also dropped.

The `traffic-manager` will still accept mux-tunnel connections but only
for the purpose of version exchange. This is needed to retain backward
compatibility with client and `traffic-agent` versions < 2.5.0.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>