Add application protocol detection to the traffic agent installer
Application protocols such as "http", "http2", "grpc", can be set in the
service port's `appProtocol` attribute in Kubernetes versions >= 1.19.
An educated guess about that protocol can be made by looking at the port
name in case the `appProtocol` is missing (a strategy used by service
meshes like Istio).

This commit checks if the application protocol can be determined for the
service that is selected for intercepts by applying a selected strategy,
and if so, assigns it to the traffic agent's environment variable
_TEL_AGENT_APP_PROTO.

The environment variable allows traffic-agents that perform L7-based
routing to select the correct protocol without probing.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>