Add mount test to smoke tests
Newer macos executors (in circle and github actions) are unable to
install macfuse successfully, so the mounts don't work in our
integration tests for macos. While we should figure out a way to make
that work, I think it's mostly because it's increasingly difficult to
install kernel extensions on new OSes for Mac.

And since we do test volume mounts in linux and windows, I think it's a
better use of our time to add a simple mount check in smoke tests (which
we run on mac) and we can come back to figuring this out when we have
spare time.

Signed-off-by: Donny Yung <donaldyung@datawire.io>