Replace --http-protocol flag with boolean --http-plaintext
It doesn't make much sense to try and override the protocol that was
selected for communication between the traffic-agent and the container
that it acts as a sidecar for when doing an intercept. The only thing
that does make sense is to disable TLS and that only makes sense when
the "getambassador.io/inject-originating-tls-secret" is used.

This commit adds the boolean `--http-plaintext` to do just that, and
removes the `--http-protocol` that were added in a previous commit.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>