More distinct session lifecycle logic
Adds the two channels connectCh and connectErr to the service and moves
the creation and run of the session into the "session" go-routine.

This ensures that session cancellation is consistent regardless of which
of the parent contexts it is that get cancelled.

Thanks @josecv for the ideas on how to improve this.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>