Make the mgrProxy mutable again and add a ManagerProxy interface
We want to maintain one grpc connection between from the root daemon
to the user daemon, and that connection will expose the traffic
manager API through the ManagerProxy with a lifecycle of the daemon.
The actual manager client that it proxies has a short session lifecycle,
so it must be possible to replace it.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>