Simplify creation of the TrafficManager and treat it like a Session
This commit copies how the root daemon manages a session. The session
in the user-daemon is the former `TrafficManager` struct. The creation
of the session is simplified so that the actual instance is created
only when everything it contains has been set up. This removes the
need for a lot of synchronization using channels.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>