Make root daemon Connect return DaemonStatus instead of Empty
In case the user daemon would crash or otherwise fail to do send a
proper `Disconnect` to the root daemon, the root daemon might have an
older session running. When that happens, the user daemon must detect
this during a `Connedt`, send a `Disconnect` and the `Connect` again.
The root daemon `Connect` must therefore return the `DaemonStatus`

Signed-off-by: Thomas Hallgren <thomas@datawire.io>