Don't start daemons to list commands
The root daemon is never needed in order to produce a command listing
and starting the user-daemon shouldn't happen just because the user
types `telepresence status`. This commit instead detects that the
user daemon isn't running (will become a rare condition) and prints
some info about that when the list of available commands is printed.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>