Drop global command flags for kubernetes and networking
Moves handling of kubernetes flags to the commands that use them
(currently `connect`, `clusterid`, `genyaml`, and `legacy`) and the
networking flags to `connect`.

Parsing of the flags is retained so that this change doesn't break
existing scripts but a warning is printed on the console if the flags
are used.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>