Use an object oriented pattern to eliminate long type switches
Adds the `k8sapi.Object` interface with methods for refresing, updating,
and deleting. Also add an `k8sapi.Workload` extending that interface
with methods to retrieve the pod template, the number of replicas and
to answer if the object has been updated since some generation.

The new types are implemented for the resource kinds `Service`, `Pod`,
`Deployment`, `ReplicaSet`, and `StatefulSet`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>