Fix problems with rotating log file on Windows
The RotatingFile on Windows was flawed. It didn't append when not
rotating and it would always rotate when using daily rotation because
the creation time of the file never changed (when renaming a file in
Windows, the source file receives the timestamp of the file that gets
overridden, go figure).

Signed-off-by: Thomas Hallgren <thomas@datawire.io>