Deprecate --http-match in favor of --http-header
The flag `--http-match` is renamed to `--http-header` to avoid confusion
as the new `--http-path-xxx` flags are introduced. A command line like:

    telepresence intercept --http-match=all --http-path-prefix=/api ...

didn't really make sense. Is it matching all requests, or just the
ones prefix with '/api'? Using `--http-header` makes it clearer:

    telepresence intercept --http-header=all --http-path-prefix=/api ...

The old flag can still be used but will render a deprecation warning
that suggests using the new flag. It no longer shows up in the help.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>