Give kubeception longer to respond
While we kept on retrying for 10 minutes, we never gave kubeception
longer than 10 seconds to respond. It's not unlikely that this might
have triggered a large number of create requests (one was sent every
10 second) that got cancelled before they had a chance to complete.

This commit changes the request timeout from 10 seconds to 5 minutes
which will limit the number of retries from 60 to 2.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>