Add proper logging to the telepresence CLI process
The CLI process was utilizing context managed logging (dlog). Yet, all
calls that risked output something to the logs (such as metriton
reporting) used a `WithDiscardingLogger` context which resulted in the
loss of some rather critical info such as failure to send metriton
reports.

This commit makes the CLI log to a proper logfile named `cli.log` which
is rotated daily, and removes all occurrences of `WithDiscardingLogger`.
The `cli.log` is included in the logs gathered by the `gather-logs`
command, only a specific daemon log isn't chosen and only if it isn't
empty.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>