Don't hide errors from WithStartedConnector
An error returned from `getRemoteCommands` was never propagated. It was
assumed that it would indicate that no user daemon was running. This
commit ensures that other errors are propagated correctly. It also
changes the logic slightly so that the call to `getRemoteCommands` is
prevented in case the user just wants to do `telepresence quit`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>