Remove the TCP recursion blocker
The recursion blocker for TCP traffic was too intrusive and caused
timeouts and is therefore removed in this commit. Recursions that may
occur when the cluster has access to the workstation's network need to
be dealt with using other network control mechanisms such as
`never_proxy` or by configuring container network so that it cannot
reach the host network.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>