Check in the telepresence-chart.tgz in the git repo
This file is embedded and needs to be included in the vendor directory
in downstream components. It must be included in the repository.

This commit changes `make build` so that it generates a temporary
`telepresence-chart.tgz` used for that build only (and hence embedded).
Once embedded, `git restore` is called to get rid of the changes. A new
`make build-version` target is added that doesn't restore the file.
The `build-version` target is used when pushing a new executable or
when doing `prepare-release`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>