Ensure that telepresence-chart.tgz isn't git-modified on each build
Changes `make build` so that it generates a telepresence-chart.tgz that
is used for that build only (and hence embedded) and then git restored.
A new `make build-version` target is added that doesn't restore the
file. This target is used when pushing a new executable.

The `telepresence-chart.tgz` is also regenerated and committed by the
`prepare-release` target.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>