rpc/userdaemon: Sync with saas_app.git
It's a problem for differing definitions of the same types.  It wasn't a
problem before because before the "telepresence.userdaemon" protobuf
package only existed in saas_app.  But it exists here now too.

If y'all didn't want to have to have these definitions in OSS then they
shouldn't have shared the "telepresence.userdaemon" protobuf package name.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>