W.I.P. Make client sessions survive for a configurable amount of time
A client session must survive typical scenarios like closing the lid
of the laptop for the day and commence the work the day after.

This commit hard-codes the client session TTL to 24 hours and the agent
session TTL to 15 seconds. It also ensures that teh connector's session
terminates rather than just producing and endless sequence of "session
not found" messages in the connector.log when the session has expired.

TODO:
The session TTLs should be configurable using a traffic manager env
vars (durations) such as `CLIENT_SESSION_TTL` and `AGENT_SESSION_TTL`,
and ultimately using corresponding Helm chart settings.

Perhaps do something nicer when the session expires, such as restart
the connector automatically.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>