Let WatchAgentsNS watch only active namespaces
An active namespace in this context, is a namespace that the user has
done some explicit action on such as `list` or `intercept`. The client
has no reason to watch agents in any other namespace. Limiting the
watch this way will decrease network traffic significantly on large
clusters when the user doesn't `connect` using `--mapped-namespaces`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>