Make initial DNS recursion check more resilient to slow DNS update
When the TUN-device starts and is configured with a new DNS IP and
search paths, it takes several seconds before DNS quires gets routed
to that IP on a Windows box. This delay made the initial DNS recursion
check fail and in a state where recursion was falsely considered, which
in turn meant that queries to for the same name that arrived faster than
they could be resolved by the cluster, would yield a NXNAME response.

This commit changes the logic so that the recursion check is retried
until it actually hits the resolver.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>