Limit namespaced traffic-manager RBAC to managed namespaces
The `ServiceAccount` used by a traffic-manager that is confined to a set
of namespaces must only be a `RoleBinding`-subject in those namespaces.
No `ClusterRoleBinding`, nor associated `ClusterRole`, is needed and
must not be created by the Helm chart.

Removing the cluster-wide access means that nodes become unavailable.
Consequently, the pod-CIDRs must be retrieved by watching pods, and
the watcher(s) must be limited to the managed namespaces.

Another consequence is that the cluster-id cannot be set to the UID of
the "default" namespace because the traffic-manager will typically not
have access to it. Therefore, the cluster-id will fall back to the UID
of the namespace where the traffic-manager is installed.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>