Ensure that the destination folder for the enhanced free client exists
A flow starting with the `telepresence login` after installing the
client would not create the directory where the enhanced free client was
installed before attempting to download it. This resulted in "error
installing updated enhanced free client: unable to create temporary
file in ...".

This commit ensures that the directory is present before the download
attempt is made.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>