Add an emptyDir mount to the traffic-agent and use it to export mounts
This commit adds an `emptyDir` volume mount to the `traffic-agent`
which can be written to regardless of what security context the pod
is using. The volume is mounted under `/tel_app_exports`. Each
app container will have a directory entry there, and under that
directory, all mounts will be symlinked to their original location.

The `/tel_app_exports/<container-name>` is then what's exported by the
`sftp` server during an intercept involving that container.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>