Remove processes parented by an intercept when the intercept is removed
## The problem
An intercept can be started with a `--` flag to spawn a command that
inherits the environment of the intercepted process. E.g.
```
$ telepresence intercept mysvc -- localsvc
```
The command would start the `localsvc` binary and not terminate until
that process terminates. This became a problem when a `telepresence
leave mysvc` was issued from another terminal. The intercept ceased to
exist but the `localsvc` binary kept on running, unaware that its
"parent" no longer was active.

## The solution
The `user-daemon` has a new gRPC call `AddInterceptor` that registers
the relationship between an intercept ID and a process PID. A
corresponding `RemoveInterceptor` exists to remove the relationship.
When the CLI process (`telepresence`) creates a parented interceptor, it
uses the `AddInterceptor` method to register a relationship between the
current intercept ID and its own IP.

Whenever the `user-daemon` is about to remove an intercept (as the
result of a call to `leave` or `quit`), it now checks if there's a
relationship between the ID of that intercept and a PID. The process is
terminated if such a relationship is found.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>