Fix so that mount-point is reported correctly on updated intercepts
An intercept with preview enabled calls the traffic-manager's
`UpdateIntercept` method. It returns an `InterceptInfo` without a
client mount point (of course, the traffic-manager has no idea where
the client mounts things). As a result, the output shown to the
user as a response to using `telpresence intercept` would not include
the mount point. A subsequent `telepresence list` would though.

This commit ensures that the mount point returned by the user-daemon
in the original call to `CreateIntercept` is preserved and reassigned
to the `InterceptInfo` after an update. The commit also contains some
minor refactoring to get rid of duplicated code in the user-daemon
when obtaining the mount point.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>