Fix problem with multiple kubeconfig files in KUBECONFIG env
Since the user daemon is long-running, the client would convert a
setting of the `KUBECONFIG` environment variable to a `--kubeconfig`
flag. This works well in most cases but fails when the `KUBECONFIG`
contains several path entries, because the `--kubeconfig` flag will only
accept one path.

This commit changes so that both the `--kubeconfig` flag and the
`KUBECONFIG` environment are set to the connector daemon verbatim. The
daemon then sets or unsets its os environment accordingly.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>