Modify the WithKubeConfigExtension test utility to use multiple files
This commit changes the `WithKubeConfigExtension` test support function
so that instead of copying the current kubeconfig file to another file
which is then used, it creates an additional file with a context named
`extra`, a copy of the cluster which contains the extension named
`extra`, and then adds that file as a path entry to the `KUBECONFIG`
environment variable. The tests using this function can then select
whether to use the extension by passing the flag `--context extra` when
invoking `telepresence connect`.

The main reason for this change was to make sure that multiple entries
in the `KUBECONFIG` environment variable works OK.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>