Fix agent-injector misbehaviour when owner is an unsupported workload.
Some workloads have owners of unsupported type. One example is Argo's
`Rollout`. The agent-injector doesn't handle this correctly. Instead
of generating a `ConfigMap` entry for the controlled workload
(typically a`ReplicaSet`), it attempts to use the `Rollout` and fails.
The correct action must be to actually use the `ReplicaSet` and don't
look further for an owner.

This commit ensures that objets controlled by a `Rollout` are
considered unowned with respect to Telepresence notion of an owner.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>