Ensure intercept uniqueness based on agent port number
When Telepresence discovers possible interceptable routes from services
to workloads, it may discover several service ports that target the
exact same container port. It will then add multiple intercepts to
the agent configuration that uses the same agent port.

The service name or service ID is used when identifying the route to
use when intercepting, but are irrelevant for other use-cases, and
since it's an error if rest of the intercept config isn't be equal when
the agent port is equal, the slice of intercepts can be normalized based
on the agent port.

This commit provides the `PortUniqueIntercepts` method that returns
the agent-port unique list of intercepts and ensures that it is used
where relevant.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>