Make traffic-agents configuration follow suit when things change
A service that is removed or added will often affect the set of possible
routes that can be intercepted. This commit adds a watcher for services
and logic to discover what existing entries that may be affected by
the service addition/modification/deletion.

In addition, the mutating webhook now also listens for pod deletions, so
that the corresponding configmap entry can be removed when the owner
workload cease to exist.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>