structured output streaming (#2602)
* Make traffic-agents configuration follow suit when things change

A service that is removed or added will often affect the set of possible
routes that can be intercepted. This commit adds a watcher for services
and logic to discover what existing entries that may be affected by
the service addition/modification/deletion.

In addition, the mutating webhook now also listens for pod deletions, so
that the corresponding configmap entry can be removed when the owner
workload cease to exist.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Make intercept state know about all possible intercepts

An intercept state must know about all possible intercepts (i.e. the
intercepts using the same port) in order to be matchable against an
intercept specification.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Update changelog.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Fix dlog.Warn -> dlog.Warnf

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Fix problem with watcher not terminating correctly when cancelled

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Fix problem uninstall --agent nnn or --all-agents

A cluster-wide watcher would be started when no namespace was given.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* chart

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Fix problem with config being recreated by webhook delete event.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Ensure proper cleanup in integration_tests

The other-NNN namespace and the Helm install in the custom-NNN namespace
wasn't cleaned up properly.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Make daemon session terminate properly when traffic-manager dies

A Helm uninstall of the traffic-manager left the root daemon in a bad
state, constantly trying to reconnect to obtain cluster info. This
commit ensures that the session dies.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Handle relative mount points

The `telepresence intercept --mount <mountPoint>` now handles relative
paths on non-windows platforms. Windows will still require the path to
be a drive letter followed by a colon.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>

* Refactor intercept cleanup logic to be done on removeIntercept.

It really made no sense to do it in a separate GC goroutine when the
user's session may already be terminated and it may be no longer be
possible to use the user's credentials to contact systema for cleanup.

Signed-off-by: Jose Cortes <josecortes@datawire.io>

* Don't store intercept API keys in a separate map.

Doesn't really make sense to since intercept termination will happen
with the client's credentials.

Signed-off-by: Jose Cortes <josecortes@datawire.io>

* Return codes.NotFound when no intercept is found

Co-authored-by: Thomas Hallgren <thomas@datawire.io>
Signed-off-by: Jose Cortes <josecortes@datawire.io>

* Address PR feedback

Signed-off-by: Jose Cortes <josecortes@datawire.io>

* add structured streaming to pkg output

Signed-off-by: Raphael Reyna <raphaelreyna@protonmail.com>

* Add StructuredStreamer unit test

Signed-off-by: Raphael Reyna <raphaelreyna@protonmail.com>

* Add more coverage around StrcutredStreamer

Signed-off-by: Raphael Reyna <raphaelreyna@protonmail.com>

Co-authored-by: Thomas Hallgren <thomas@datawire.io>
Co-authored-by: Jose Cortes <josecortes@datawire.io>