Upgrade windows dependencies
Updates wintun.dll from version 0.12 to 0.14.1.
Updates winfsp from 1.9 to 1.10

Download of wintun.dll was moved into makefile because changing the
workflow doesn't have any affect until the PR is merged and that
resulted in a mismatch between the downloaded dll and the version of
the go library in the go.mod file.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>