Deprecate config setting for images.webhookAgentImage
There is only one type of agent image since 2.6, and that is the one
injected by the webhook, so it no longer serves a purpose to have a
configuration setting for both `agentImage` and `webhookAgentImage`.
This commit deprecates the latter. If set, a warning is logged and the
value ends up in the `agentImage`.

This commit also changes the `images.webhookRegistry` so that it
defaults the setting of `images.registry` which in turn defaults to
the environment variable `TELEPRESENCE_REGISTRY`. Prior to this change,
the `images.webhookRegistry` would default directly to the environment
variable.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>