Restore the _TEL_AGENT_NAME to the traffic-agent's set of env vars
Some customers use this environment variable to inject the TLS secret
annotations and give it a unique name although the configuration is
static. That works since telepresence always expand those annotations
with the current set of environment variables. Or well, it did work
until `_TEL_AGENT_NAME` was removed. That's why it is restored in this
commit.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>