Add InjectPolicy to control when traffic-agents are injected
Prior to this change, the mutating webhook would inject the
`traffic-agent` container into created pods whenever an attempt was made
to intercept a workload where the container was missing. Setting the
`telepresence.getambassador.io/inject-traffic-agent` to `disabled` would
prevent the injection.

Now, a new `InjectionPolicy` is introduced that makes it possible to
control this behavior further. The policy can have one of the following
settings:

- OnDemand: is the old behavior and also the default setting.
- OnDemandWhenEnabled: is similar to the old behavior, but it requires
  that `telepresence.getambassador.io/inject-traffic-agent` is set to
  `enabled`
- WhenEnabled: will inject immediately when the pod is created or
  updated. It will not wait for an attempt to intercept.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>