Add InjectPolicy to control when traffic-agents are injected
Prior to this change, the mutating webhook would inject the
`traffic-agent` container into created pods whenever an attempt was made
to intercept a workload where the container was missing. Setting the
`telepresence.getambassador.io/inject-traffic-agent` to `disabled` would
prevent the injection.

Now, a new `InjectionPolicy` is introduced that makes it possible to
control this behavior further:

- OnDemand: is the old behavior and also the default setting.
- WhenEnabled: if the annotation is enabled, then inject in advance
  when the pod is created or updated. Nothing gets injected unless the
  annotation is present.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>