Only consider last session when counting the number of errors in the log
The log can grow quite large over a day and contain a lot of sessions
but only the errors from the last one are relevant when reporting the
number of errors found after a command has been issued. If that command
(or the current session) has zero errors, then there's really no point
in telling the user to go have a look in the logs in case it fails.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>