Let the telepresence client cache the session-info
Prior to this change, the client would forget its session-id when it
was forcefully disconnected and reconnecting to the traffic-manager
would create a new session. As a result, the old session was lost and
all intercepts belonging to it was unavailable. The traffic-manager
would however keep the session alive for 24 hours, and during that time
the intercepts would remain active. An active tcp intercept effectively
blocks all other intercepts to the same workload.

This commit stores the session-id in the client's cache and removes it
only after a successful `Depart`. If the session-id is found in the
cache when the client connects, then the client will use that in an
attempted call to `Remain`. If the call succeeds, the client reuse the
session and not call `ArriveAsClient` to create a new session.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>