Introduce protocol in PreparedIntercept and a new PortIdentifier type
The `PreparedIntercept` struct will now always contain the selected
protocol for the intercepted port.

The `PortIdentifier` is a string type intended for the
`servicePortIdentifier` field of the `InterceptSpec` The new type was
motivated by increasing complexity of this field which now can be
either a name or a number, and be combined with a protocol.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>