Fix tagging problem in Makefile
The `TELEPRESENCE_VERSION` was evaluated late (as always when using
`=` or `?=`) so using `make build push-image` didn't work. The final
push referred to a different version then the one tagged by the build,
because `./build-aux/genversion` got called twice.

This commit pegs the `TELEPRESENCE_VERSION` to the one generated the
first time it is generated.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>