Handle pods with no volumes and automountServiceAccountToken == false
A workload with `automountServiceAccountToken: false` that has no
volumes, will have `/spec/volumes` unset. The webhook injector would
always assume that `/spec/volumes` was at least an empty list, and thus
create an incorrect patch attempting to add entries to that missing
value. The failing replicaset would show the following:
```
Error creating: Internal error occurred: add operation does not apply: doc is missing path: "/spec/volumes/-": missing value
```
This commit ensures that the patch is correct by first checking if the
value is undefined, and if so, replacing it with a list of volumes
instead of attempting to add them.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>