Move everything route related to routing package
The `routing.Route` is not directly associated with a `vif.Device`
because it has an `Interface` of its own, and passing it as a parameter
to the methods `AddStaticRoute` and `RemoveStaticRoute` on an instance
of `vif.Device` isn't strictly correct because the device is never used.

This commit moves those functions to `routing.Route` and renames them
(to avoid stutter) to `AddStatic` and `RemoveStatic`.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>