Deprecate uninstall --everything instead of removing it
The `telepresence uninstall` command was completely removed in 2.7.0
which may create some unnecessary churn for our users. The functionality
was moved in unmodified under `telepresence helm uninstall`. That was
a bit unfortunate, because `helm` isn't involved when uninstalling
agents. Also, it requires a `--everything` flag to actually remove
the traffic-manager.

This commit reinstates the `telepresence uninstall` command but
deprecates the `--everything` flag. It can still be used, but now it
delegates to `telepresence helm uninstall` and informs the user that
it's deprecated.

This commit also changes `telepresence helm uninstall` into a command
that just performs a `helm uninstall` of the traffic-manager.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>