Ensure that tunnel streams are closed when connection attempts fail
If a traffic-agent makes an attempt to connect to an interceptor process
on the workstation, and no such process exists, that results in the
dial being rejected, which is correctly propagated, but unfortunately,
the stream between the traffic-manager and the workstation is not being
closed. This commit ensures that the close happens and also adds some
additional security around canceling of streams.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>