Ensure that commands that Exec uses the command's stdout/stderr
The `proc.Run` would simply use `os.Stdin`, `os.Stdout`, and `os.Stderr`
when executing a command and provided no way of changing this.

This commit adds a `proc.Stdio` interface that implements the subset
of the `*cobra.Command` that provides access to the standard I/O and
ensures that commands that call `proc.Run` provides themselves as the
implementation of that interface.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>