Introduce an atomic sessionQuitting variable to solve potential race
When the session disconnects, it will first use a read-lock, then
cancel the session context, then release the read-lock and attempt to
take a write-lock. That's all fine as long as the session cancellation
doesn't affect recursive read-locks, as is the case when `RunCommand`
uses its `connectorService` to remove an intercept. The `RunCommand`
already holds a read-lock, which prevents the quit from taking its
write-lock, that attempt however, blocks the remove from taking another
read-lock.

The solution was to introduce an atomic boolean that gets set when the
session is quitting. When set, no attempts will be made to take a
read-lock. Instead, a `cancelled session` error is returned. Then, as
soon as the session quit procedure is complete, the settings is cleared.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>